[cmake] fix llvm out-of-tree build for one-step build strategy (#193)
The current one-step build strategy requires LLVM sources to be placed
in the buddy-mlir project and named 'llvm'. This makes packaging a
difficult task, as users can't build buddy-mlir with other working llvm source
code that has a different version to the one provided by buddy-mlir.

This commit sets 'LLVM_PROJECT_SOURCE_DIR' to be the up dir of
'LLVM_MAIN_SRC_DIR'. 'LLVM_MAIN_SRC_DIR' is set when building within the llvm
source directory, so it is safe to use it as the relative path to llvm project.

Signed-off-by: Avimitin <dev@avimit.in>